requirements.txt
discord.py>=2.2.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=1.0.0